BHL/services/sqs_consumer/requirements.txt
boto3==1.28.44
requests==2.31.0
python-dotenv==1.0.0 